requirements.txt
asgiref==3.2.7
attrs==19.3.0
autobahn==20.4.2
Automat==20.2.0
beautifulsoup4==4.9.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==4.0.0
click==7.1.2
constantly==15.1.0
cryptography==2.9.1
cycler==0.10.0
Django==3.0.3
django-appconf==1.0.4
django-bootstrap3==12.0.3
django-bootstrap4==1.1.1
django-braces==1.14.0
django-imagekit==2.0.4
django-multiselectfield==0.1.12
google-play-scraper==0.2.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
joblib==1.0.1
kiwisolver==1.3.1
lxml==4.6.3
matplotlib==3.4.2
#mysql-connector-python==8.0.24
#mysqlclient==2.0.3
nltk==3.6.2
numpy==1.20.2
oauthlib==3.1.0
pilkit==2.0
Pillow==8.2.0
play-scraper==0.6.0
protobuf==3.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
regex==2021.4.4
requests==2.25.1
requests-futures==1.0.0
requests-oauthlib==1.3.0
service-identity==18.1.0
simplejson==3.17.2
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
textblob==0.15.3
tqdm==4.60.0
tweepy==3.10.0
txaio==20.4.1
urllib3==1.25.9
wincertstore==0.2
zope.interface==5.1.0
